chore(bangumi): update api generate

+++ bangumi-api/src/main/kotlin/moe/kurenai/bangumi/models/UserAccessToken.kt
@@ -41,7 +41,7 @@ data class UserAccessToken(
     val refreshToken: kotlin.String,
 
     @SerialName(value = "user_id")
-    val userId: kotlin.Long
+    val userId: kotlin.Int
 
 ) {
